Fix #70: Modularize output.rs into submodules (#80)
Split the 1,300+ line output.rs into smaller, focused modules:
- output/mod.rs - Re-exports and shared types
- output/config.rs - OutputConfig struct
- output/utils.rs - Shared utility functions (wrap_text, continuation_prefix, etc.)
- output/tree.rs - TreeFormatter for buffered output
- output/streaming.rs - StreamingFormatter for console streaming
- output/markdown.rs - MarkdownFormatter for markdown output
- output/json.rs - JSON output function

This improves maintainability, testability, and makes the code easier to navigate.
Each module now has a single responsibility.
